requirements.txt
streamlit
langgraph
langchain_core
langchain-text-splitters
langchain-community
langchain-openai